Move things around to make code base easier to read
