+++ reconcile/templating/lib/rendering.py
